integrated pydantic for type checking in requests and refactored
