+++ functions/requirements.txt
@@ -4,3 +4,6 @@ google-cloud-texttospeech
 python-dotenv
 partialjson
 mutagen
+fastapi
+pydantic
+uvicorn
